requirements.txt
# Eye Tracking Application - Cross-platform requirements
# Compatible with Windows, macOS, and Linux

# Core Computer Vision Libraries
opencv-python>=4.8.0
mediapipe>=0.10.13
numpy>=1.24.0

# Data Analysis (optional - used for advanced features)
scikit-learn>=1.3.0
matplotlib>=3.7.0
fpdf2==2.8.4
streamlit==1.48.0
streamlit-webrtc==0.63.4
av==14.4.0
# Platform-specific notes:
# Windows: Visual C++ 14.0+ required for some packages
# macOS: No additional requirements
# Linux: libGL1-mesa-glx package may be required for OpenCV